server: set up a web server
This is the initial commit for Indecision App build up while following
The Complete React Web Developer Course (with Redux) course [0].

This commit adds a README.md file with a list of packages installed in order
to get a web server set up. It also adds a boilerplate html file to get
displayed when the server starts.

Ref.: Lecture 7

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>