app.js: render updated results to DOM
Up until now the Count attribute did not change whatsoever upon the pressing
of buttons.

This commit implements the logic for the buttons behaviors as well as wire
them up to the Count value.

Also, the calls for rendering the values to the screen are being set up properly,
resulting in the buttons changing and the Count value being displayed accordingly.

Ref.: Lecture 17

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>