plauyground: add build-it-visible app
Build-it-visible is a small app built from scratch with the knowledge already
gained through the course so far and is intended to live in the playground
directory.

The app displays a button that shows and hides information. I added extra
information on the steps involved in the making of the app.

Ref.: Leture 21

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>