app.js: render elements of options array
This commit changes the way the Options component renders on the screen. It
was printing a paragraph with a static text and the Option component inside,
without managing contents of an array of options.

This commit implements a JavaSCript call for each element of the options array
variable and renders it on the screen.

Ref.: Lecture 28

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>