playground: rebuild counter with current knowledge
This commit introduces a basic rebuild of the counter app using the tecniques
learned so far up until lecture 32.

It was calling renderCounterApp() manually at every method, and now it is using
the React Component tool to get the same effect. It also implements the binding
of 'this' that will be useful in future implementations.

Note that for the bindings on the constructor method, there is no code commited.
Therefore, revisit lecture 30 to understand method binding.

As this is a part of a bigger set of commits, some older code is left out as
comments down below the file for future references.

Ref.: Lecture 32

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>