playground: add setState to automatically render data
The method setState() is responsible to manipulate the state of the object
and to automatically render it on the screen. It receives an old state
'prevState' as an argument and is able to alter it and then re-render it.

In order to get that behavior, a 'state' object has to be created inside
the constructor with the default data and the attributes of the state object,
along with their data.

Ref.: Lecture 33

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>