app.js: add methods and reverse data flow
This commit introduces functions being passed as props to methods, so
buttons can alter objects of the app, such as 'options'. This way the data
flow is being reversed, allowing a child to communicate with the parent by
calling the method.

The methods being added with this commit are 'handleDeleteOptions' that
is passed to Options and removes all the elements of the 'options' array.

Also, 'handlePick' is being implemented with the logic of the src/playground/
jsx-indecision-app.js, to pick one of the elements of the array and display it
as an alert on the browser screen

Ref.: Lecture 36

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>