app.js: wire up AddOptions to revert data flow
This commit implements a parent 'handleAddOption' method, responsible at this
moment to log whatever the user inputs in the AddOptions form element. This
data is sent to the (parent) method by the child 'AddOption' which in turn
is responsible to get the data from the input text box.

This code makes sure the basic functionality of the method is working.

Ref.: Lecture 37

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>