app.js: add data into options array
This commit implements the pieces to pass data into the options array, defined
in the constructor method. This is accomplished by making use of functions as
props to get data flow both downstream (from parents to children) as upstream
(from children up to parents).

Also add some validation to the AddOption method, making sure the data passed
by user is not duplicate nor an empty string. By doing this, error messages
make sure to display that to the user.

Reference for the concat method:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/concat

Ref.: Lecture 37

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>