app.js: make classes into stateless functional components
Stateless functional components can be created from those classes that do not
depend on the current or previous application state.

They are faster and lighter than class based components, and also more readable
than the latter, plus they do not have the overhead of all the class code and
the load to extend React. Components, which ends up bringing useless features to
this scope.

They are also easier to test but they do not handle 'this', and values can be
passed into the functions via props.

Ref.: Lecture 40

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>