app.js: simplify this.setState function sintaxes
In ES6 sintax, return functions will implicitly be called when there is only
one element returned.

It happens that when the return element is an object inside a function, a pair
of parenthesis is needed in order to make JavaScript understand the object is
not a function body, but instead, an object.

Say we want to return { options: [] }, we need to pass it to:

	this.setState( () => {} );

Double curly braces are not an option, but wrap them into parenthesis is. Like so:

	this.setState( () => ({}) );

JavaScript ES6 sintax will understand the above as a functions implicitly
returning an object.

Ref.: Lecture 43

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>