app.js: add button to remove a singular option
When the option text is passed to AddOption in order to be rendered on the
screen, a new button 'remove' will be passed along with it. This button
is currently console.logging a string and the event information on the console.

This is a commit to implement the functionality and the final logic will be
coming up next in this same patchset (branch).

Ref.: Lecture 43

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>