app.js: introduce Lifecycle Methods
This commit introduces three lifecycle methods and puts them in place to use.
Lifecycle methods are called when the components they track change state. The
three methods being added are componentDidMount, componentDidUpdate and
componentWillUnmount, and they will be explored in the future.

Documentation about them is found here [0]

[0] https://reactjs.org/docs/react-component.html

Ref.: Lecture 44

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>