app.js: make 'handleDeleteOptions' permanent
By now, all the changes made in the 'options' array were stored in the
localStorage environment variable so to preserv the current state of the array,
except for the 'handleDeleteOptions' method: all data would go back on screen
as soon as the page reloaded, ignoring that the Remove All button was just
pressed.

The code being added here makes the 'handleDeleteOptions' permanent, preventing
the removed elements to be restored when the page is refreshed.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>