app.js: add two tweaks about usability
The input field is currently not being wiped when we add an option, making
us to select and delete the previous contents of the box to add another one.
The first tweak added with this commit is deleting the input text when we
press 'Add Option'.

The second tweak is adding a paragraph with a short usage instruction when
there is no items in the 'option' array.

Ref.: Lecture 45

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>