infra: implement webpack bare bones structure
This commit adds webpack as a local module and create a script
to run it from the CLI in packages.json. As part of the webpack basic
implementation, a config file 'webpack.config.js' was also put in place,
in order to record the project's pointers to webpack (such as entry point
and output file).

The public/scripts/app.js was removed as it will no longer be required.

Also the .gitignore file was modified not to track the public/bundle.js
(the new output file) as it is automatically generated by webpack and will
be changing every single commit.

The pointers for React and ReactDOM were removed from index.html and will be
handled by webpack in future implementations.

Lastly, the old code for app.js was moved to playground for future reference,
while a new one will be generated to make use of webpack features.

Documention for the new dependencies and configuration is found at [0]
and [1] below.

[0] https://webpack.js.org
[1] https://nodejs.org/api/path.html#path_path_join_paths

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>