app.js: import npm modules
This commit provides the basic structure for importing npm modules to the code.
In app.js, the modules imported are react[1] and react-dom[2], but validator
version 8.0.0[0] were also installed and added to packages.json.

React.createElement() is being used because support for JSX (Babel)
was not yet introduced.

Ref.: Lecture 53

[0] https://www.npmjs.com/package/validator
[1] https://www.npmjs.com/package/react
[2] https://www.npmjs.com/package/react

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>