app.js: set up Babel with webpack
By default, webpack do not understand JSX without a configuration setup
to teach it what to do. To accomplish that we use babel-loader that
allows customization of webpack when it loads a given file.

So this commit adds babel-core to the project, which is similar to babel-cli,
but needs configuration to run properly.

Also add babel-loader, which is a webpack plugin to teach webpack how to run
babel when webpack is provided certain files.

The actual changes in code start at webpack.config.js, by creating a module
object[0] to configure webpack to use babel. In order to pass the 'env' and
'react' presets, an extra .babelrvc configuration file is added.

[0] https://webpack.js.org/configuration/module/#module-rules

Ref.: Lecture 54

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>