app.js: refactor for usability and maintainability
While restoring the functional app.js that was previously moved to playground,
it is a good practice to break it down into a file per component, and solve
the respective imports altogether.

This commit accomplishes this refactoring, improving the code in clarity as
well as in maintainability.

Note that for Options and Option, the link between the two was related to
Options. So Option was exported and imported on Options, and Options was in
turn exported and imported into IndecisionApp.

Be aware that the diff will not show the code taken from app.js, since the
code is being extracted from playground/app.js

Ref.: Lecture 55

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>