webpack.conf: add source-maps into webpack
When there is an exception or error in the application we are building,
the logs complain and point to a line from bundle.js, which is the only
running file in our application (and is created by babel).

Sometimes it is difficult to identify where the error is coming from, in
the ES6 code we are build.

This commit adds source-maps to the project, a developer tool that points
to an ES6 code from our original code. The production one will be inttroduced
in the future.

Documentation can be found at [0]

[0] https://webpack.js.org/configuration/devtool

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>