code: refactor for class-properties
This commit moves the class-based components to a new class-properties syntax,
making use of the newly installed babel-plugin-transform-class-properties

The transform-class-properties plugin lets us refactor code of a class by
removing the need for a constructor for the class instantiation. Instead,
the variables are declared without type, straight from the class declaration.

Besides using class-properties to set up default state values of the components,
by removing the constructors, we use them to clear the binding issue we have been
facing with event handlers so far and have been solving with a manual binding
inside the constructor.

The reason is that arrow functions do not bind 'this' to themselves, and
instead, are bound to the class 'this', so the binding becomes unbreakable.

So, the atual changes are:

 - set the default state out of the constructor
 - turn event handler functions into arrow functions, to solve the binding issue.
 - update README.md to register the new settings.

Ref.: Lecture 58

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>