indecision-app: set up react modal
One way of passing JSX contents to a component is via prop, like:

  ReactDOM.render(<IndecisionApp content={template}/>, document.getElementById('app'));

Another way is via child technique:

  ReactDOM.render((
    <Layout>
        // contents of the JSX
    </Layout>
  ), document.getElementById('app'));

This commit sets up react modal[0], which passes JSX contents using the 'child'
technique.

Modals bring the popup box that is triggered when handlePick handler is
activated (when 'What should I do?' button is pressed).

This patch adds a new component file src/components/OptionModal.js that controls
the popup box, while in indecisionApp.js a new handleClearSelectedOption handler
is created to set the selectedOption as undefined when the modal box is closed.

[0] https://github.com/reactjs/react-modal

Ref.: Lectures 60 and 61

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>