app.js: configure base style
The style of the application will use scss, and some steps are required to
accomplish the configuration.

This commit implements a basic configuration of a simple .css file, in order
to get everything set up to configure SCSS.

The README.md file has instructions to create this new initial setup.

Ref.: Lecture 64

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>