app.js: introduce scss structure
This commit introduces the scss loaders and file. The concept is similar
to JSX, in which we write code that is not recognizable by the browser
and call some loaders (from the webpack.config.js) file to turn them
browser readable.

Also we need to import the converter (node-scss module) so to get the
scss syntax into regular css.

README.md is also populated with respective instructions.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>