css: set up font family and size
The scss styles are being implemented in separate files called partials.
In our case, the _base.scss file is a partial of styles.scss and contains
the basic setup for the fonts of our app at this stage.

The html tag setting up font-size at 62.5% is a converstion constant
between pixels and rems (1rem = 16px), and setting it up at 62.5% makes
the conversion more intuitive, since 0.625*16px=10px, meaning our values
get closer to pixels to be assigned to variables.

Example:
  at 62.5%, 1rem = 10px;
  1.6rem = 16px;
  2.2rem = 22px;
... and so on

Ref.: Lecture 65

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>