styles: apply styles to header block
The chosen style concept for the application uses style breakdown, called
partials, and the BEM convention, that stands for Block Element Modifier.

The partials are imported into the entry styles.scss file, which in turn is
loaded as the main style file.

The Header partial implements classNames following the BEM convention and
reads 'header__title', for example.

Also, in order to be applied, the elements of the Header.js component need to
properly indicate what class they are part of.

Note that JSX does not support the standard 'class=' attribute, but does
'className=' instead.

[0] www.getbem.com

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>