modal: add transition to modal box
This commit adds an opacity transition to the modal box, making it fade in
and fade out when functioning.

For documentation purposes, the '>' sign at the first class in _modal.scss means
that only the divs that are sons of 'ReactModalPortal' will be affected, and not
their descendants.

Ref.: Lecture 71

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>