modal: style the modal text
This commit formats the modal interface text and look. Classes were created
and applied to the modal box to make it more friendly.

Also a minor issue with word wrapping was also addressed here, making words
to break after they are longer than the box. This fix was also extended to
the options list menu, which had the same behavior (words by default do not
break).

Ref.: Lecture 71

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>